pre-train core via reconstruction, crop image based on retinotopy (#15)
* resnet

* update

* update

* option to save evaluation result to yaml file

* clean evaluation function

* pretrain by reconstruction

* update

* del loss image outputs

* use item

* update

* update

* update

* update

* update

* code cleaning

* update

* update

* remove l1 reg calcualtion in validation

* stacked2d core

* update

* use ssim for reconstruction

* update

* update

* update

* update

* fix regularization calculation. add ds_scale to scale loss by the size of the dataset

* iterator for train ds

* use correlation as early stopping metric

* use correlation as early stopping metric

* use correlation as early stopping metric

* update

* pretrain script only save core module

* update

* update

* update

* update

* update

* update

* update

* correlation as objective

* update

* combine LR scheduler with early stopping

* update

* lr scheduler max_reduce instead of es_patience

* util function to load checkpoint

* neuron-wise correlation loss

* update

* update

* fix scheduler update of num_reduce. fix utils.evaluate printing

* simplify reconstruction

* update

* recon mse

* recon mse

* update

* pretrain model missing reg_scale

* plot_image_response same y-axis

* add norm_mode==1 to normalize responses. remove count mistake in utils.inference.

* update

* update

* update

* update

* pretrain recon add one more conv

* use grid predictor. norm_mode use population min and max instead of individual neurons

* scale imagenet validation image to 255.0 before normalization to match sensorium images

* grid_sampler input_dim=2

* scale image argument

* dynamic patching in ViT

* crop image based on retinotopy

* update

* code cleaning

* code cleaning for ViT

* update

* add --grid_predictor_dim argument to set number of dimension of coordinate grid

* submission load entire model

* crop pretrain

* pretrain plot

* recon shape

* reconstruction readout

* add dropout to stacked

+++ src/sensorium/metrics.py
@@ -9,7 +9,7 @@ def correlation(
     y1: t.Union[torch.Tensor, np.ndarray],
     y2: t.Union[torch.Tensor, np.ndarray],
     axis: t.Union[None, int, t.Tuple[int]] = -1,
-    eps: int = 1e-8,
+    eps: float = 1e-8,
     **kwargs
 ):
     """
